Modernize tests by following Intellij tips (#300)
* Modernize tests by following Intellij tips

* rollback import *

* rollback import * 2

* use proper import

* Replace Optional Guava by Optional Java (#301)

* Replace Optional Guava by Optional Java

* rollback import *

* organize imports

* Modernize tests by following Intellij tips

* rollback import *

* rollback import * 2

* use proper import

* rebase

+++ src/test/java/com/urbanairship/api/push/model/notification/ios/IOSDevicePayloadTest.java
@@ -111,7 +111,7 @@ public class IOSDevicePayloadTest {
         Boolean hidden = true;
         assertEquals(hidden, m.getMediaAttachment().get().getOptions().get().getHidden().get());
         assertEquals("unique ID", m.getThreadId().get());
-        assertEquals(0.5, m.getRelevanceScore().get().doubleValue(), 0.0f);
+        assertEquals(0.5, m.getRelevanceScore().get(), 0.0f);
     }
 
     @Test(expected = IllegalArgumentException.class)
